+++ src/FileManagement.h
@@ -5,7 +5,7 @@
 namespace FileManagement {
     
     //template<typename T>
-    int WriteBufferToFile(std::string path, const char* buffer, int bufferSize);
-    char* LoadBufferFromFile(std::string path, int* bufferSize);
+    int WriteBufferToFile(std::string path, const char* buffer, uint64_t bufferSize);
+    char* LoadBufferFromFile(std::string path, uint64_t* bufferSize);
     void CompareFiles(std::string firstPath, std::string secondPath);
 }
